Moved server related stuff to the server folder.
